backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
networkx==3.2.1
feedparser==6.0.11
python-dotenv==1.0.1
httpx==0.27.2
rapidfuzz==3.10.0
